Implement searching though zips for images.
Added new project files since iws is no longer supported
Added new setting option to setup a temp folder where an image in a zip
will be copied to.  This folder should be empty, as currently
implemented it will remove the contents of the folder every time it
copies a new image to be displayed.
ImageHandler completely reworked.